ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4
